% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean}
\alias{mean}
\title{Function for calculating mean and rounding on three decimals}
\usage{
mean(values)
}
\arguments{
\item{values}{A vector with values}
}
\value{
The mean rounded on three decimals
}
\description{
Function for calculating mean and rounding on three decimals
}
