% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{students}
\alias{students}
\title{Information of students}
\format{An object of class \code{data.frame} with 8 rows and 5 columns.}
\usage{
students
}
\description{
The datatset contains the heights, weigths, names, gender and the age of the students
}
\keyword{datasets}
